Code refactoring
Removed unused code.
Moved the style folder to the top level.
Refactoring the <Loader /> component. This component
has become universal.
