+++ src/Components/MainPage.js
@@ -1,7 +1,7 @@
 import React from "react";
 import { Button } from "react-bootstrap";
 import { Link } from "react-router-dom";
-import "./styles/main.css";
+import "../styles/main.css";
 
 const MainPage = () => {
   return (
