+++ src/Components/authorization/signUp/FormSignUp.js
@@ -5,7 +5,7 @@ import { useHistory } from "react-router";
 import { setUserCredential, setLoading } from "../../../redux/actions";
 import { useDispatch, useSelector } from "react-redux";
 import { spinner } from "../../../redux/selectors";
-import { Loader } from "../../Loader";
+import Loader from "../../Loader";
 
 const FormSignUp = () => {
   const history = useHistory();
@@ -89,7 +89,7 @@ const FormSignUp = () => {
     <div className="pageForm_sign-up">
       {loadingPage ? (
         <div className="formSingUp_loader">
-          <Loader />
+          <Loader animation="border" variant="success" />
         </div>
       ) : (
         <div>
