+++ src/Components/todoApp/DisplayTodos.js
@@ -1,10 +1,10 @@
 import React from "react";
 import ButtonRemove from "./buttonsControl/ButtonRemove";
-import CompletedTodo from "./checkboxesComplete/CompletedTodo";
+import CompletedTodo from "./checkboxesComplete/CheckboxTodo";
 import TodoItem from "./TodoItem";
 import { spinner } from "../../redux/selectors";
 import { useSelector } from "react-redux";
-import { Loader } from "../Loader";
+import Loader from "../Loader";
 
 const DisplayTodos = ({ data, refetch }) => {
   const arrTodos = data?.todo;
@@ -19,8 +19,8 @@ const DisplayTodos = ({ data, refetch }) => {
           <ButtonRemove id={id} refetch={refetch} />
         </div>
       ))}
-      
-      {loading && <Loader />}
+
+      {loading && <Loader animation="border" variant="success" />}
     </div>
   );
 };
